refacto : 워커 프로세스 분리
